// src/containers/screens/Home/Escort/RecordScreen/types.ts
export enum EscortState {
  Started = 'Started',
  Finished = 'Finished',
}

export enum RecordType {
  Video = 'Video',
  Audio = 'Audio',
}

export enum PanicMode {
  Active = 'Active',
  Inactive = 'Inactive',
}